requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.2.0
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.4.0
braceexpand==0.1.7
certifi==2022.5.18.1
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
commonmark==0.9.1
discord.py @ git+https://github.com/Rapptz/discord.py.git
dnspython==2.2.1
frozenlist==1.3.0
idna==2.10
import-expression==1.1.4
jishaku==2.5.0
motor==3.0.0
multidict==6.0.2
numpy==1.22.4
pandas==1.4.2
pycparser==2.21
Pygments==2.12.0
pymongo==4.1.1
PyNaCl==1.5.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.25.1
rich==12.4.1
six==1.16.0
typing_extensions==4.2.0
urllib3==1.26.9
wavelink==1.2.5
yarl==1.7.2
youtube-related==1.0.3
